[IMP] hr_timesheet/saas~17.5: update rules due to `message_partner_ids`
Following https://github.com/odoo/odoo/pull/172130 and
https://github.com/odoo/enterprise/pull/66163, some ir.rule domains
that are `noupdate=1` needs to be updated.

The PR introduced a new non-stored field `message_partner_ids`
with the idea of replacing domains that looks for `message_partner_ids`
on the `task_id` and `project_id` with a `search` field, that is
faster to use than relying on multiple sub-queries.

closes odoo/upgrade#6260

Related: odoo/odoo#172130
Related: odoo/enterprise#66163
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>