[IMP] account/16.0: fail if tax report can't be migrated
opw-3884262

As explained in the code: Odoo 16 does not support multiple report lines with
the same `code`, even if associated to different reports.

This is to detect such a scenario early and fail with an actionable error
message, hopefully preventing customers from opening tickets to ask about that.

Part-of: odoo/upgrade#6275
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>