[FIX] hr_expense/0.0.0: set sheet approval_state from its current state
Commit odoo/odoo@f79ff3fb3746 introduced a new `approval_state` field
which is used as dependency for the `state` compute.

So when the `approval_state` is NULL (default) the `state` might
unexpectedly revert to `draft` when recomputing it.

For ex. consider an upgrade to saas~17.3, a expense sheet which is
already 'Approved' (state = 'approved') before upgrade, once upgrade
(before this commit) the `approval_state` will be falsy and trying to
do anything that recompute the `state` field (for ex. creating the
expense journal entry using the `Post Journal Entries` button)
will reset it to `draft`.

Note: The issue is espectially visible since `saas~17.3` as the
 `approval_state` is now checked first.