[FIX] base/0.0.0: handle deletion of fields selection
Since saas~13.3, the deletion of `ir.model.fields.selection` records
expect the linked model to exists in the registry[^1].

This fail in builds[^2] of PRs that adds a new model with selection
fields in stable branch (tolerated in granita branches).

Explicitly delete the records linked to an ignored model or field.

Also automatically ignore all fields of an ignored model.

[^1]: https://github.com/odoo/odoo/blob/98d36b80ed1b4f11491301d3b7b5818d81edf542/odoo/addons/base/models/ir_model.py#L1357
[^2]: https://web.archive.org/web/20240718113452/https://runbot.odoo.com/runbot/build/65630424

closes odoo/upgrade#6292

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>