[IMP] project,sale,_*/saas~17.5: remove and update fields/views
This commit removes some fields and views removed but also migrate some
data in the new fields created in the related PRs. The idea behind the
changes made in the related PRs is to have a link on project for differents
models (MO, PO, SO, stock picking and BoM) to easy link those documents
to a specific project. Moreover, the `analytic_distribution` field in MO
and BoM will be replaced by the project field and inside
`project.project` model, the mixin (`analytic.plan.fields.mixin`) is used
to add a column per plan to be able to define a default analytic
distribution to the SOLs and POLs once the project is linked to SO and PO
for instance. It means the data in analytic distribution of MOs and BoMs
has to be migrate to those new fields inside project. The idea implemented
for the migration is we will only generate a project for the MO and BoM
containing a analytic distribution with only a line at 100 percent, we
will no longer cover a complex analytic distribution on those models
(MO, BoM).

This commit also renames `analytic_account_id` field into `account_id` in
`project.project` model because `analytic.plan.fields.mixin` uses
`account_id` field and not `analytic_account_id` field. It also migrates
the analytic account set on the projects to the column representing the
plan set on that analytic account. For instance, if the analytic account
of the project is linked to `Departments` plan and not the `Projects`
plan then the analytic account will be set on `account_id` to continue
to be able to timesheets inside that project and `x_plan{plan_id}_id`
column to be coherent with the `plan_id` field in
`account.analytic.account` model.

See https://github.com/odoo/odoo/pull/169868
See https://github.com/odoo/enterprise/pull/65054

task-3985258

closes odoo/upgrade#6296

Related: odoo/odoo#169868
Related: odoo/enterprise#65054
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>