[FIX] website/17.0.1.0: activate/create a cowed view
The database upgraded from a lower version to version 17, If a website has a
header menu layout of "Hamburger menu" then a list of menus is not visible to
the customer.

Due to 'no_autohide_menu' view is deactivated(In case, If the customer has once
used the header layout of 'Sidebar' then 'no_autohide_menu' is created for that
website and after unused of that header the view is deactivated. because that
view is used in that header layout) or unavailable in views as it was not
required in the previous version for that header.
From v17, 'template_header_hamburger' header layout should require an active
"no_autohide_menu" view for that website.

To solve this issue, I have created a cowed view for that particular website and
for the already created cowed view, just activate them.

opw - 4054201
Req - 1840005

closes odoo/upgrade#6294

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>