[FIX] base/0.0.0: deprecation of `to_text`
Since upgrade still has to deal with Python 2, create a severely cut
down version as it's only used once: just decode from `bytes` or let
the content pass through. That way on P2 we decode the `str` to
`unicode` (assuming that's what `re` returns) and on P3 it's a no-op.

Related to odoo/odoo#173995

closes odoo/upgrade#6299

Related: odoo/enterprise#67081
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>