[IMP] pos_restaurant/saas~17.5.1.0: Table number instead of name
In restaurant table, the 'name' field become an integer field
called table_number. The change made in this commit aims at conserving
previous name that were integer. For the name that were not, the field
is set to a unique integer higher than the previous maximum table number
in the database.

Community PR: odoo/odoo#173108
Enterprise PR: odoo/enterprise#67152

closes odoo/upgrade#6302

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>