[FIX] base/0.0.0: fix views in non English languages
Steps to reproduce:
1. Create a DB with other than 'en_US' language.
2. Edit view a through studio add a field that have been remove or
   renamed.
3. Upgrade to next version.

The view will be disabled while it should have been fixed.

Root issue: if error message here
https://github.com/odoo/upgrade/blob/5861a34817d7665bad8614001662654f3573a325/migrations/base/0.0.0/pre-models-ir_ui_view.py#L717
is in a language other than English, then `heuristic_fixes`
https://github.com/odoo/upgrade/blob/5861a34817d7665bad8614001662654f3573a325/migrations/base/0.0.0/pre-models-ir_ui_view.py#L389
is unable to fix the view because it won't match the patterns.

In this patch we ensure that the logs are always in English for the
checks in the views fixer.

closes odoo/upgrade#6312

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>