[FIX] spreadsheet_edition/saas~17.2: handle broken revision history
On version `17.2` new foreign key relation is introduced[^1] between `id`
and `parent_revision_id` on model **spreadhseet.revision**.

Also new UNIQUE INDEXES are introduced[^2] on that model to ensure having
at most single child per revision and at most single parent per revision.
So it is expected to have 1 straight line when we look at the revision history
while it can be broken easily. You can delete any of them by enabling dev mode
and going to:
Settings --> Technical --> Spreadsheet --> Revisions

So if there are any parentless (except the head) revision it would be assumed
as the head and the upgrade will set NULL to the parent_revision_id here[^3]

Steps to reproduce:
1. Install spreadsheet_edition in `17.0`
2. Create spreadsheet document and make several changes
3. Open new tab and with dev mode go to:
   Settings --> Technical --> Spreadsheet --> Revisions
   then remove revision from middle
4. Leave from document on first tab and try to upgrade to `17.2`

As the upgrade will set NULL to the parent_revision_id of
head and the child of the revision you have removed.
You will get traceback like this when it is trying to create UNIQUE indexes:

```
  File "/home/odoo/src/enterprise/saas-17.2/spreadsheet_edition/models/spreadsheet_revision.py", line 32, in init
    self.env.cr.execute(
  File "/home/odoo/src/odoo/saas-17.2/odoo/sql_db.py", line 346, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.UniqueViolation: could not create unique index "spreadsheet_revision_initial_unique"
DETAIL:  Key (res_model, res_id)=(documents.document, 14) is duplicated.

```
`Snapshot`

For not reading 1000s of revisions on each time when the user opens a document,
we are using snapshotting. Basically on each snapshot new revision with command
type `SNAPSHOT_CREATED` is created and all revisions are set active=False
and the last state of document is stored in filestore. So when any user reads
document we read all newer revisions after the last snapshot
on top of last snapshot.
Snapshotting can be triggered by these ways:
1. When all of the users which opened the same document leaves the document,
   it takes snapshot when the last one leaves
2. Every user can go to developer mode and trigger snapshotting
   manually by `File -> Snapshot`
3. If the document didn't receieve any change during last 2 hour it will take
   snapshot no matter if there is user which keeps that document opened.

Basically the revision history could be broken in several ways. During the
upgrade we have to ensure to have 1 straight line in history for each document.

If there are 4 revisions A--->B--->C--->D, assume somebody removed
the C with dev mode. There are 2 types of users:
1. Users which the documents was opened when the removal happened.
2. Users which opened the document after the removal.

`Possible snapshotting scenarios`
1. If second type of user opens document he/she will not see C and D
   as they were removed before he/she opens the document. At the moment
   he/she makes any change it will be notified to all first and second
   type users, so C and D will disappear from first type users too.
   After this point no matter how the snapshotting happens, there is no way
   to retrieve the C and D.
2. If there is no any second type of user opens the document after removal
   and one of the first type of user keeps making changes which they will
   be written on top of D. At the end if he/she takes snapshot at the end,
   new changes will not disappear. It will take snapshot of the document state
   for that first type user.

In conclusion, if snapshot occurs there is no way to retrieve the broken
history before snapshot. We can not create phantom revisions for connecting the
gaps because as I explained above, when C is removed, C--->D branch is broken
and there might be new revisions which are written on top of B.

There are mainly 2 broken scenarios:
1. There is gap and snapshot.
2. There is gap but no snapshot.

In this commit I made a fix for the case 1.
If there is gap and snapshot, I collect the revisions starting from
last snapshot and continue tracking the history until we reach
next break (if there is one) and remove all revisions which are not on
that path `LAST_SNAPSHOT--->....--->FIRST_GAP_AFTER_LAST_SNAPSHOT`

[^1]: https://github.com/odoo/enterprise/blob/db5036588f5cc4d3554e03ed27e65c8f2cb9e42a/spreadsheet_edition/models/spreadsheet_revision.py#L23
[^2]: https://github.com/odoo/enterprise/blob/db5036588f5cc4d3554e03ed27e65c8f2cb9e42a/spreadsheet_edition/models/spreadsheet_revision.py#L31-L49
[^3]: https://github.com/odoo/upgrade/blob/294a889b5f049be8e9a85dac65dc0bca1144be0f/migrations/spreadsheet_edition/saas~17.2.1.0/pre-migrate.py#L91-L109

closes odoo/upgrade#6340

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>