[IMP] documents_spreadsheet/saas~17.5: migrate documents spreadsheet after the document refactor
Purpose
=======
The documents application has been refactored (share and folder have been removed, etc),
and so we need to adapt documents_spreadsheet.

Frozen spreadsheets
===================
Before, we could freeze a spreadsheet, and share the frozen version of the spreadsheet.
It was a bit weird, because if an internal user opens the link, he was redirected to the
real spreadsheet and not to the frozen one (and so, he saw a different document if he
was logged in).

Now, everybody will see the same document (as long as they can access it). To freeze a
spreadsheet, we duplicate the document, and so we can choose to share the "live"
spreadsheet, or a frozen version (this is done with 2 different action).

We also had a model `documents.shared.spreadsheet` to store the frozen spreadsheets,
this is not needed anymore because we store the information on the documents.

We migrate those `documents.shared.spreadsheet` only if there's only one spreadsheet,
since the access token can not be shared.

Default folder
==============
There's a default folder for spreadsheet created outside of the documents application.

Users can only see their own spreadsheets in that folder.

Task-4014621
COM odoo/odoo/pull/180264
ENT odoo/enterprise/pull/65655

Part-of: odoo/upgrade#6226
Related: odoo/odoo#180264
Related: odoo/enterprise#65655
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>