[IMP] point_of_sale/saas~17.5: product_template improvements
In the related PRs, we are introducing the pos_urban_piper module and we made
few fixes and improvements in the related modules.

We renamed the description_self_order field in the product.template
model to public_description and moved it from the pos_self_order
module to the point_of_sale module.

task-4040542

Related PR:

closes odoo/upgrade#6327

Enterprise: odoo/enterprise#66733
Community: odoo/odoo#173300
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>
Co-authored-by: paso-odoo <paso@odoo.com>
Co-authored-by: shag-odoo <shag@odoo.com>