[FIX] account/0.0.0: consider fiscal position accounts in anglo_saxon
Before this patch, Anglo-Saxon accounting was evaluated based on the accounts
present in `ir_property` for stock, income, and expenses defined on the product
categories. There are cases where a fiscal position is set on the moves, and
account mapping has been created in the fiscal position for the accounts
defined on the product categories. As a result, the account set on the
Anglo-Saxon move lines will be replaced by the account used in the fiscal
position. Since this replacement account is not present in `ir_property`,
it is not considered, and the Anglo-Saxon lines with the replacement account
are not evaluated as Anglo-Saxon lines. In version 16, these lines will be
displayed as product instead of cogs.

Steps to reproduce :-

- Create a fiscal position and add an account mapping.
- Use an account from the product category and set the replacement account to
  one that is not used in the product category.
- Create an Anglo-Saxon invoice with the created fiscal position.
- Upgrade to version 16.

After upgrading, the lines with the replacement account will be displayed
as `product` lines instead of `cogs`.

closes odoo/upgrade#6328

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>