[FIX] mrp_workorder/saas~16.4: compute employee_id
Since odoo/enterprise@d4f65146 (16.0) there is a new field 'employee_id' in
'mrp.workcenter.productivity'. Later in odoo/enterprise@d5faf8653 (16.0) the
field was made a computed stored one.And then also changed in
odoo/enterprise@dd5c31331 (17.0).
In version 16, the compute method relies also on the boolean field
'allow_employee',which was introduced with a default value false. Consequently,
this condition of compute method will not be satisfied. So it can't be computed
during the upgrade process from version 15 to 16.

Additionally, we should consider records where the 'employee_id' field is null
without computed in those dbs which are already in version 16 .
So it needs to update in version 17.

opw-4051305
upg-1826818,1868458

closes odoo/upgrade#6291

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>