[IMP] account/saas~17.5.1.2 autopost of bills
We have a new field `is_manually_modified` in `account_move` model.
When importing a bill from a vendor, if this field is `False` for
the last 3 bills of this partner, it means that the bill can be
automatically validated. By default, this field is `False` and
set to `True` when the user manually modifies the bill.

However, when upgrading, all the bills will have this field set to
`False`, so the feature will be automatically activated after one
invoice instead of 3 unmodified ones.

To avoid this, we are setting the field to `True` for all past
acount.move's so that the counter stars back from 0 again.

Community PR: https://github.com/odoo/odoo/pull/174540
Enterprise PR: https://github.com/odoo/enterprise/pull/67788

closes odoo/upgrade#6356

Task-id: 3958958
Related: odoo/odoo#174540
Related: odoo/enterprise#67788
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>