[IMP] l10n_es_edi_*/saas~17.5.1.1: remove tbai dependency to account_edi
The TicketBAI edi is no longer using the `account.edi` framework and its
`account.edi.document` model, in favor of a new model
`l10n_es_edi_tbai.document`.

The logic has to be migrated.
The `l10n_es_tbai_chain_index` and the xmls sent were previously stored
on the `account.move`, they are now stored on the tbai document.
The tbai documents are created from the xmls stored on the move and
move datas.
This is not enough to distiguish an xml who was sent successfully and an
xml where there was a request error.  This information is taken
from the `account.edi.document` with a `blocking_level='warning'.
The failed xmls were not stored on the db, then not migrated.

The PR alse removes the dependency to the `l10n_es_edi_sii` module.
It causes the `l10n_es_edi.certificate` model to be seperated in
`l10n_es_edi_sii.certificate` and `l10n_es_edi_tbai.certificate`,
which are pretty much the same.

closes odoo/upgrade#6373

Pr: odoo/odoo#166183
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>