[FIX] payment/0.0.0: defuse check required fields
Some fields are required for specific providers. The check
was correctly defused for saas~16.5 only, but any force update
of record for any other place than saas~16.5 will lead to failure
in case false as clients would have to fix their providers anyways
and this also happens when test databases are neutralized.

[TBG-1424](https://upgrade.odoo.com/web#id=1424&cids=1&menu_id=107&action=178&model=upgrade.request.traceback.group&view_type=form)

closes odoo/upgrade#6387

See: #5710
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>