[IMP] account_reports/saas~17.5: Linking account_move to report
This commit aims to remove the field tax_closing_end_date.
A new field tax_closing_report_id is the way to know if it is a closing move.
With that change all old closing moves need to fill this new field.
We fill that field with the default report for that company country.

Also removed a deprecated field on account_move (`tax_report_control_error`)
See odoo/enterprise#68551

closes odoo/upgrade#6390

Related: odoo/odoo#180095
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>