[IMP] mrp_mps/saas~17.5: remove manufacturing_period_to_display
This PR remove `manufacturing_period_to_display` from `res.company` and
`res.config.settings` and replaces it with 4 new integer fields that will
fulfil the same functions:
- manufacturing_period_to_display_year
- manufacturing_period_to_display_month
- manufacturing_period_to_display_week
- manufacturing_period_to_display_day

Thus, the original value it contains needs to be moved to the field
corresponding to the setting selected in `manufacturing_period`.
We don't include 'year' in the script since it is an entirely new option.

see odoo/enterprise#68548

closes odoo/upgrade#6406

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>