[IMP] account/saas~13.4: speedup bank statement line refactor
OPW-3884369
UPG-1775402

Although on a single db, a `chunk_size` of 3000[^1] was found to speed up the overall
processing time of the bank statement lines by almost 2x wrt to 1000 (~8h vs ~13.5h).

The db, this was tested on, is relatively large:

```SQL
pied@(none):pied_1775402> SELECT COUNT(*) FROM account_bank_statement_line
+---------+
| count   |
|---------|
| 9579838 |
+---------+
```

Also, mark as `UNLOGGED` tables, used to perform the refactoring and dropped before
finishing the upgrade.

[^1]: 2000 and 4000 were also tested and yielded very close results. However, the test
with `chunk_size`=3000 was performed during a period of much higher load for the test
server.

Part-of: odoo/upgrade#6418
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>