[FIX] repair/saas~16.4: Fix repair notes without taxes
This query adds the information from the Operations tab, removed in the new
version, to the Notes tab,, so users without Invoice access rights can still
see that information.

If no taxes exist in the m2m with `repair.fee`, because of the JOIN clause,
the query will retrieve nothing and that information will not show up.
By making it optional, it will show up even if not taxes are present.
The other JOIN clauses are on mandatory fields.

opw-4138489

closes odoo/upgrade#6465

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>