[IMP] spreadsheet_dashboard_edition/saas~17.5: rename file name field
This commit removes the field `file_name`. Now the field `spreadsheet_file_name`
is used.

It has been promoted and moved to `spreadsheet.mixin` because it needs to be
used in the `quality_control` module.

It can't be renamed propertly because of "the bug of the mixins".
See the warning message when trying to rename the field:

```
The field spreadsheet.dashboard.spreadsheet_file_name_custom is deleted during the upgrade.
This might be due to the bug of the mixins, not marking the fields on their rightful module.
Since this is not a stored field,
we let it be deleted, it will be created back by the ORM on the first
 `-u` on the according module if needed, hopefully marked with the correct module this time
```

odoo/odoo#180275
odoo/enterprise#70077

closes odoo/upgrade#6511

Task: 4187954
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>