[IMP] repair/saas~17.5: add stock.rule
When the manufacturing app is installed, it adds an MTO (replenish on
order) rule. We would like this rule to also exist when only repair is
installed, so for consistency to the code change in the community PR, we
ensure that existing DBs with only repair installed will have the MTO
rule added.

COM PR: https://github.com/odoo/odoo/pull/168348

closes odoo/upgrade#6576

Task: 3935079
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>