[IMP] sale_subscription/saas~18.1: remove sale order alert
This commit relates to the 'base_automation' module dependency removal
from 'sale_subscription' as from now on all the automation will be done
in Studio. As a result, we are dropping the 'sale.order.alert' model
and its derived views. In addition, we are also removing the 'health'
field from sale orders and its reports, as it will be no longer used.

task: 4181465
See odoo/enterprise#70180

closes odoo/upgrade#6527

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>