[IMP] appointment/saas~17.5: add a field for the appointment status
We now use the new field `appointment_status` to better manage the
status of appointments.
Remove the `appointment_attended` field as it's now a specific status
of the new field (attended).
While at it, we also allow the manual confirmation option for staff user.
This leads to the renaming of the field `resource_manual_confirmation`
into `appointment_manual_confirmation`.

ENT PR: odoo/enterprise#66076

Task-4028357

closes odoo/upgrade#6214

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: amdi-odoo amdi@odoo.com
Co-authored-by: jeh-odoo jeh@odoo.com