[IMP] documents_hr/saas~17.5: migrate documents_hr after share-pocalyse
Purpose
=======
Following the document share-pocalyse refactoring, we migrate the folder
document_hr_folder from <documents.folder> to <documents.document>.

The HR folder was configured to allow internal users to view their own
documents (through groups and user_specific) and to allow the hr user to edit
all documents. The main migration will take care of adding view access to
document owner and other accesses will be managed by adding access to the hr
folder.

Migrate the "archived user share"
=================================
When we archive a user, we can create a share, so he can still access its HR related
documents. This commit migrates those shares.

Because we can have many shares for the same employee, and also because we don't store
the tokens, we use the temporary table `documents.redirect`, that is already used to
migrate some shares. The model will be removed in the future, when we think it's ok to
not support the old URLs.

Task-4014621
COM odoo/odoo/pull/180264
ENT odoo/enterprise/pull/65655

Part-of: odoo/upgrade#6226
Related: odoo/odoo#180264
Related: odoo/enterprise#65655
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: Pierre-Yves Dufays <pydu@odoo.com>