[FIX] account/0.0.0: ensure deferred accounts exist
Some charts of account modules set deferred accounts on the
company in a post init hook, and if the accounts are not found
this results in a warning. When this happens during the upgrade
on the CI it flags a failure, so the solution is to ensure the
needed accounts are loaded before they are used on the company.

closes odoo/upgrade#6563

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>