[FIX] product/saas~17.4: markup not correctly set up
When upgrading, customers complain that the display name is not correct,
This is due to the fact that a previous commit odoo/odoo#159746
a new price_markup field was added but it was not populated based
on existing pricelists.
The use of price_markup is to set price_discount to -price_discount
So by definition we need to set price_markup to -price_discount
if the base price is cost or standard_price (technical name).

display_applied_on is also created and populated based on existing values

odoo/odoo#181391

closes odoo/upgrade#6570

Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>