[FIX] tools/matt.py: run all tests in 12 series
Since this patch odoo/odoo@483cc20b (Odoo 13) we no longer rely on
`__init__.py` to load tests. Unfortunately for Odoo 12 and saas~12.3 we
still do.

Since the split of upgrade-util we copy the scripts and merge the
folders. This causes the init file to be overridden. Hence some tests
are not run.

Same has been done on the platform with odoo/upgrade-platform#680.

closes odoo/upgrade#6572

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>