[IMP] *: Using class name corresponding to odoo models
If the class does not have the correct name, it is possible to force the
model name using `_name`.
In order to minimize changes for migration scripts, the solution of
forcing the `_name` is made instead of correcting the class names.

Related PRs:
odoo/odoo#178200
odoo/enterprise#69762
odoo/documentation#11115
odoo/upgrade-util#143

closes odoo/upgrade#6577

Related: odoo/design-themes#988
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>