[FIX] account_accountant_batch_payment/saas~18.1: remove pay reject
The wizard has been removed just before the official release but kept to
avoid any issues. It is being removed now.

It was removed because there is no unfold of the batch in the
reconciliation widget, and all it was doing should now be done from the
form view of the batch instead.

See:
odoo/odoo#182470
odoo/enterprise#71288

closes odoo/upgrade#6586

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>