[FIX] base_import_module/0.0.0: set default module_type for new modules
In v17, a new
[selection](https://github.com/odoo/odoo/commit/92565aae27274ee12fa691b5832050eef18a4e17)
field `module_type`, was introduced to differentiate
between `official` and `industries` modules. By default, this field is set to
`official`. However, during migration, when any new module is discovered,
the default value is not updated for that new module because the field already
exists in databases being migrated from `v17.0`.

Steps to reproduce:
- Create a database in `v17.0` with just the base module.
- Migrate the database to `v17.4`.
- In the migrated database, go to Apps and search for any new module
  introduced in `v17.4` (e.g., `marketing_card`).
- Try to open the module info for this module in the Kanban view.

closes odoo/upgrade#6589

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>