[FIX] base/0.0.0: Ensure default view for ir_act_window_view mapping
Remove records from ir_act_window_view that don't have
a view_id or a default view
(except this views 'activity','form','graph','pivot' ,'kanban','tree'),
as mapping an action without
a view is pointless. This also creates issues when users
try to access other views related to the action.

For example, if a customer creates a view mapping for
action(456) on the project.task model:

        |action_id | view_mode | view_id
 ________________________________________
1)      |456       | form      | 345
2)      |456       | tree      | null
3)      |456       | grid      | null

If no view exists with type='grid' and model='project.task',
this will raise a validation error for other views.

closes odoo/upgrade#6615

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>