[IMP] base/0.0.0: Updated Turkey to Türkiye
This is a followup PR of this one: https://github.com/odoo/upgrade/pull/6558
Community PR: https://github.com/odoo/odoo/pull/180758
Enterprise PR: https://github.com/odoo/enterprise/pull/70304
It accomodates for the DBs in 18 for the country name update.
Before this commit, DBs in v18.0, with the old name of the Turkish republic,
when upgraded to higher versions would not have the country name upgraded.

task-4182931

closes odoo/upgrade#6610

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>