[IMP] documents/saas~17.5: robustify the share create activity migration
We convert in python the following that was previously done using a sql query:

In the migration only one share is migrated to a document folder. The other
are migrated as simple alias which we name secondary alias.

The share (and now the document folder) supports the creation of an activity
when an attachment is received by email through an alias. So for the main alias
we just copy the activity fields to the folder that replace the share/alias.

In order to keep that feature for the secondary aliases, we set in the
alias_defaults the values of the activity creation fields of the share
(activity_option, ..., prefixed with "create_") so that the document "template"
created when receiving a mail through the alias has the information to create
the activity on the documents created from all the attachments of the email
(using the templates). Indeed, that first created document (template) is
archived right away and only serve as a template to create the real documents.
So we can use the create_xx of that document to create the activities on the
final document holding the attachment of the email.

To sum up, we migrate the main share activity create fields to the folder that
replace that share, and we migrate the activity create fields of the secondary
share in the alias_defaults of the corresponding secondary alias. We don't set
alias_defaults activity fields as it is taken in charge by the document model
(and will be very out of sync as soon as modified on the document).

Task-4014621

closes odoo/upgrade#6619

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>