[FIX] sign/0.0.0: Add normalized email to demo partner
For sign requests, the signer email is taken from the partner normalized email,
and it's required to not be empty. The demo data uses `base.partner_demo_portal`
as partner.

If a database is marked as demo but it has deleted the records, they will be
created sequentally by the modules. Since `res.partner.normalized_email` is
defined in the `mail` module, it will be empty when the record is loaded from
base, and it will trigger the constraint later on when the demo data from sign
is loaded.

tbg-1362

```
File "/home/odoo/src/enterprise/16.0/sign/models/sign_request.py", line 761, in _check_signer_email_validity
    raise ValidationError(_("All signers must have valid email addresses"))
odoo.exceptions.ValidationError: All signers must have valid email addresses
```

closes odoo/upgrade#6627

Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>