[FIX] website/17.0.1.0: set a cowed view in inherit for related website
The database upgraded from a lower version to version 17, If a website has a
header menu layout of "Hamburger menu" then a list of menus is not visible to
the customer.

Because 'no_autohide_menu' views inherit from the 'Main layout' view and during
upgrade we create a cowed view for 'no_autohide_menu' view by copying from
related standard view also inherited view is the standard view of 'Main Layout'
but if a cowed view is already available of 'Main Layout' in the customer
database then that needs to be set as an inherited view for a newly created cowd
view of 'no_autohide_menu' to resolve this issue.

opw - 4246249
Req - 2114055

closes odoo/upgrade#6651

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>