[IMP] documents/saas~17.5: migrate workflow rule (sharepocalypse)
Workflow rule have been discarded in v18 and replaced by server action. This
migration convert the existing workflow rule (except the one referenced by a
xmlid as they are already converted in the "data") into server actions.
See odoo/enterprise#65655

The workflow rule are converted in server action with multiple children that
execute part of the workflow rule (mark activities as done, add an activity,
add tags, ...). The only thing that we can't convert is the criterion:
- condition_type
- domain
- criteria_partner_id
- criteria_owner_id
- required_tag_ids
- excluded_tag_ids
- limited_to_single_record
As we can't convert those condition, we only pin the action to the folder if
there are no extra criterion.

We also don't migrate the "note" as there is no equivalent in the server
action (See odoo/enterprise#73376).

Task-4283330

closes odoo/upgrade#6729

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>