[ADD] account_iso20022/saas~18.1: move iso20022_charge_bearer
In the related commit, we move the ISO 20022 ChrgBr field from the
`account.batch.payment` model to the `account.payment` model to allow
defining it more granularly.

In this upgrade we copy the values from the batch payment to the
payments it contains and remove the field on batch payments.

See https://github.com/odoo/enterprise/pull/72972

task-4275038

closes odoo/upgrade#6688

Related: odoo/enterprise#72972
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>