[FIX] documents/saas~17.5: point documents_facet to documents_document
As you know, `documents.folder` is being moved to `documents.document`[^1]
and the all records pointing to folder is being pointed to document.

There is unique constraint `documents_facet_name_unique`[^2] on fields
**folder_id and name** on model `documents.facet`, so it is causing issue
when the records from `documents.facet` are being pointed to
`documents.document` as there might occur conflict for that unique
constraint on certain pointing ordering.

For example, if we have
1. Three documents **A, B, and C** with id 1, 2, and 3 respectively.
2. Two folders **D and E** with id 3 and 4 respectively.
3. Two facets **F and G** with same name `X` and folder_id 3 and 4 respectively

When we move the folders to `documents.document` model, think D gets id=4
and E gets id=5. Now after pointing the facets F and G to `documents.document`
we would expect the folder_id of F and G be 4 and 5 respectively.

But the problem is at the moment when we point the F to new model, because
at that time we will have 2 facets with same name and folder_id before the
re-pointing the facet G.

In order to fix the issue I have moved the removal of `documents.facet` to pre
script from end script as it is going to be removed anyway.

UPG-2167844
TBG-1574

```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/18.0/odoo/service/server.py", line 1305, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "<decorator-gen-13>", line 2, in new
  File "/home/odoo/src/odoo/18.0/odoo/tools/func.py", line 97, in locked
    return func(inst, *args, **kwargs)
  File "/home/odoo/src/odoo/18.0/odoo/modules/registry.py", line 131, in new
    odoo.modules.load_modules(registry, force_demo, status, update_module)
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 475, in load_modules
    processed_modules += load_marked_modules(env, graph,
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 363, in load_marked_modules
    loaded, processed = load_module_graph(
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 180, in load_module_graph
    migrations.migrate_module(package, 'pre')
  File "/home/odoo/src/odoo/18.0/odoo/modules/migration.py", line 254, in migrate_module
    mod.migrate(self.cr, installed_version)
  File "/tmp/tmp_2xbfvgz/migrations/documents/saas~17.5.1.4/pre-migrate.py", line 107, in migrate
    util.explode_execute(
  File "/tmp/tmp_2xbfvgz/migrations/util/pg.py", line 350, in explode_execute
    return parallel_execute(
  File "/tmp/tmp_2xbfvgz/migrations/util/pg.py", line 175, in parallel_execute
    return parallel_execute_impl(cr, queries, logger=_logger)
  File "/tmp/tmp_2xbfvgz/migrations/util/pg.py", line 92, in _parallel_execute_threaded
    cr.execute(queries[0])
  File "/home/odoo/src/odoo/18.0/odoo/sql_db.py", line 354, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.UniqueViolation: duplicate key value violates unique constraint "documents_facet_name_unique"
DETAIL:  Key (folder_id, name)=(10, {"en_US": "Status"}) already exists.
```

[^1]: https://github.com/odoo/enterprise/pull/65655
[^2]: https://github.com/odoo/enterprise/blob/3e8594098a47258a8aedfacf8aca4488d55e8907/documents/models/documents_facet.py#L24-L26

closes odoo/upgrade#6676

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>