[FIX] account_report/saas~18.1: revert totals below sections removal
This reverts commit 956b236ed288145453cd35b64ca11ccc5e5da93a.

We got reports that the "Totals Below Sections" setting is actually important in the US and LATAM countries. The usage is there to always display amounts like that on reports, and there is a lot of resistance to softwares not allowing that. We hence reintroduce this feature.

closes odoo/upgrade#6678

Related: https://github.com/odoo/enterprise/pull/72930
Related: odoo/enterprise#72930
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>