[FIX] documents/saas~17.5: fix the propagation of the access
Bug
===
Fix the "edit" access propagation.

During the migration of the HR documents,
the HR users could have lost their access
on the payslips, we had before:
- read group: internal user
- write group: hr user
- user_specific: True
- user_specific_write: False

After the migration, the HR users were members on the HR folder,
which is correct, but not on the payslips inside that folder
(only the employee was member).

Check that the user is in the group before adding him member, for
user specific folder (and that he's in the write group to let him
owner for user specific write).

Technical
=========
Before, we had:
```xml
<record id="documents_folder_groups_rule" model="ir.rule">
    <field name="name">Documents.folder: groups</field>
    <field name="model_id" ref="model_documents_folder"/>
    <field name="groups" eval="[(4, ref('base.group_user'))]"/>
    <field name="domain_force">[
        '|', '|',
            '&amp;',
                ('read_group_ids', '=', False),
                ('group_ids', '=', False),
            ('group_ids', 'in', [g.id for g in user.groups_id]),
            ('read_group_ids', 'in', [g.id for g in user.groups_id]),
        ]</field>
</record>

<record id="documents_document_readonly_rule" model="ir.rule">
    <field name="name">Documents.document: readonly rule</field>
    <field name="model_id" ref="model_documents_document"/>
    <field name="groups" eval="[(4, ref('base.group_user'))]"/>
    <field name="domain_force">[
        '&amp;',
            ('folder_id.read_group_ids', 'in', [g.id for g in user.groups_id]),
            '|',
                ('folder_id.user_specific', '=', False),
                ('owner_id', '=', user.id)
        ]</field>
    <field name="perm_read" eval="True"/>
    <field name="perm_write" eval="False"/>
    <field name="perm_create" eval="False"/>
    <field name="perm_unlink" eval="False"/>
</record>

<record id="documents_document_write_rule" model="ir.rule">
    <field name="name">Documents.document: folder write groups</field>
    <field name="model_id" ref="model_documents_document"/>
    <field name="groups" eval="[(4, ref('base.group_user'))]"/>
    <field name="domain_force">[
        '|',
            '&amp;',
                ('folder_id.group_ids', 'in', [g.id for g in user.groups_id]),
                '|',
                    ('folder_id.user_specific_write', '=', False),
                    ('owner_id', '=', user.id),
            '&amp;',
                ('folder_id.read_group_ids', '=', False),
                ('folder_id.group_ids', '=', False)]
            </field>
    <field name="perm_read" eval="True"/>
    <field name="perm_write" eval="True"/>
    <field name="perm_create" eval="True"/>
    <field name="perm_unlink" eval="False"/>
</record>
```

We propagated the "read groups" but not the "write groups",
but it should be propagated if `user_specific_write = False`
on the parent folder.

In the same way, we only propagate the "read groups"
if `user_specific = False`.

In the weird case where the user is in the "write group",
but not in the "read group", we still propagate the access,
because the access rules `documents_document_readonly_rule`
and `documents_document_write_rule` are "unified", and not
"intersected".

Task-4307017

Part-of: odoo/upgrade#6710
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>